:recycle: Refactor code and update available commands
This commit refactors the code to add functionality for the `/start` command,
allowing users to send a message that will be echoed back by the bot.
It also updates the available commands section in the README file to include
the new options for the `/start` command. Additionally,
the commit adds the `.env` file to the `.gitignore` list and
updates the `./kbot` command in the installation instructions to include the `start` subcommand.