:rocket: Build Docker image with multi-platform support (#1)
* 🎉 Add GitHub Actions workflow for building Docker images

Add a new GitHub Actions workflow to build and push Docker images for the project. The workflow is triggered by pushes to the `test` branch or tags that match the pattern `v*.*.*`. The workflow includes several steps to log in to the container registry, extract metadata for the Docker image, and build and push the image. Some of the steps are commented out for now, but they can be uncommented later to enable additional features.

This change improves the project's development workflow and makes it easier to build and deploy Docker images.

* :test_tube: Get environments

* Disable working dir

* :pencil2: (docker): Fix cert issue

* :construction_worker: Build Docker image with multi-platform support

This commit updates the `.github/workflows/build-image.yml` file to add the `DOCKER_BUILDX_PLATFORMS` environment variable, which specifies the target platforms for building the Docker image. The platforms include `linux/amd64`, `linux/arm64`, `darwin/amd64`, `darwin/arm64`, and `windows/amd64`.

Additionally, it updates the build and push steps in the workflow to set up QEMU and Docker Buildx for multi-platform builds. It also adds an inspect step to print information about the builder, such as name, endpoint, status, flags, platforms, tags, labels, registry, and platforms.

The Makefile is also updated to include the `docker-build` and `docker-push` targets, which build and push the Docker image with the specified target platforms.

The `README.md` file is updated with instructions on how to build and push the Docker image, including custom platforms.

The `docker/Dockerfile.template` file is added, which serves as a template for the Dockerfile used to build the image. It includes instructions to build the Go binary and copy it into the final image.

The `docker/docker.mk` file is also added, which defines variables and targets for building the Docker image.

* Added VERSION and CHANGES files, Version bump to v0.1.0

* Make tag